end-user-workspace/main.py
#!/usr/bin/env python3
"""
AI Infrastructure Platform - Main Deployment Script
=================================================
Handles the deployment and management of AI infrastructure on Proxmox.
Cursor will help you use this script and interpret its output.
"""

import os
import sys
import logging
import argparse
from pathlib import Path
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('deployment.log'),
        logging.StreamHandler(sys.stdout)
    ]
)

logger = logging.getLogger(__name__)

def deploy():
    """Deploy the AI infrastructure."""
    logger.info("🚀 Starting AI infrastructure deployment...")
    
    # Validate configuration
    if not validate_config():
        logger.error("❌ Configuration validation failed")
        logger.info("💡 Ask Cursor to help you fix the configuration")
        sys.exit(1)
    
    # Test Proxmox connection
    if not test_proxmox_connection():
        logger.error("❌ Proxmox connection failed")
        logger.info("💡 Ask Cursor to help you troubleshoot the connection")
        sys.exit(1)
    
    # Deploy infrastructure
    logger.info("📦 Deploying LXC containers...")
    # TODO: Implement container deployment
    
    logger.info("🤖 Setting up LibreChat...")
    # TODO: Implement LibreChat setup
    
    logger.info("🔌 Configuring MCP server...")
    # TODO: Implement MCP server setup
    
    logger.info("🌐 Setting up Cloudflare tunnel...")
    # TODO: Implement Cloudflare setup
    
    logger.info("✅ Deployment completed successfully!")
    logger.info("💡 Ask Cursor to help you access and manage your new AI infrastructure")

def destroy():
    """Destroy the AI infrastructure."""
    logger.info("🗑️  Starting infrastructure destruction...")
    
    # TODO: Implement infrastructure cleanup
    
    logger.info("✅ Infrastructure destroyed successfully!")

def status():
    """Check infrastructure status."""
    logger.info("📊 Checking infrastructure status...")
    
    # TODO: Implement status checking
    
    logger.info("✅ Status check completed!")

def logs():
    """Show deployment logs."""
    log_file = Path("deployment.log")
    if log_file.exists():
        with open(log_file, 'r') as f:
            print(f.read())
    else:
        logger.info("No deployment logs found.")

def test_connection():
    """Test Proxmox connection."""
    logger.info("🧪 Testing Proxmox connectivity...")
    
    # TODO: Implement actual Proxmox connection test
    
    logger.info("✅ Connection test completed!")

def validate_config():
    """Validate configuration files."""
    logger.info("🔍 Validating configuration...")
    
    required_vars = [
        'PROXMOX_HOST', 'PROXMOX_USER', 'PROXMOX_PASS',
        'NETWORK_GATEWAY', 'NETWORK_SUBNET', 'DNS_SERVERS'
    ]
    
    missing_vars = []
    for var in required_vars:
        if not os.getenv(var):
            missing_vars.append(var)
    
    if missing_vars:
        logger.error(f"Missing required environment variables: {', '.join(missing_vars)}")
        logger.info("💡 Ask Cursor to help you set up the missing configuration")
        return False
    
    logger.info("✅ Configuration validation passed!")
    return True

def test_proxmox_connection():
    """Test connection to Proxmox."""
    logger.info("🔌 Testing Proxmox connectivity...")
    
    # TODO: Implement actual Proxmox connection test
    
    logger.info("✅ Proxmox connection test passed!")
    return True

def main():
    """Main entry point."""
    parser = argparse.ArgumentParser(description='AI Infrastructure Platform')
    parser.add_argument('command', choices=['deploy', 'destroy', 'status', 'logs', 'test-connection'],
                       help='Command to execute')
    
    args = parser.parse_args()
    
    try:
        if args.command == 'deploy':
            deploy()
        elif args.command == 'destroy':
            destroy()
        elif args.command == 'status':
            status()
        elif args.command == 'logs':
            logs()
        elif args.command == 'test-connection':
            test_connection()
    except KeyboardInterrupt:
        logger.info("🛑 Operation cancelled by user")
        sys.exit(1)
    except Exception as e:
        logger.error(f"❌ Error: {e}")
        logger.info("💡 Ask Cursor to help you resolve this error")
        sys.exit(1)

if __name__ == '__main__':
    main()
