end-user-workspace/requirements.txt
# AI Infrastructure Platform Dependencies
# ======================================
# Cursor will help you install and manage these dependencies

# Core dependencies
requests>=2.31.0
pyyaml>=6.0
urllib3>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Proxmox API
proxmoxer>=1.4.0

# Cloudflare integration
cloudflare>=2.19.0

# Logging and monitoring
structlog>=23.0.0
colorama>=0.4.6

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
